+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [DOCS] clarify the naming 'unseen' by setting link to IMAP flags description in RFC (thanks to [kekscode](https://github.com/kekscode)) https://github.com/bt909/imap-mailstat-exporter/pull/57
 
+* [FEATURE] allow setting mailbox password via commandline or environment variable for one account setup https://github.com/bt909/imap-mailstat-exporter/pull/67
+
 * [CHORE] update module github.com/prometheus/common from v0.47.0 to v0.48.0 https://github.com/bt909/imap-mailstat-exporter/pull/56
 * [CHORE] update module github.com/prometheus/client_golang from v1.18.0 to v1.19.0 https://github.com/bt909/imap-mailstat-exporter/pull/60
 * [CHORE] update module github.com/prometheus/common from v0.48.0 to v0.49.0 https://github.com/bt909/imap-mailstat-exporter/pull/61